refactor: enhance component organization
- Rename ColorPickerField to ColorPicker for better consistency
- Update component imports and references across the codebase
- Consolidate instance selection logic in InstanceSelector
- Improve form handling and validation in InstanceForm
- Add color picker integration with form components
- Enhance component state management and prop handling
